+++ NEWS.md
@@ -9,3 +9,4 @@
 * enable default country selection at startup
 * introduce custom_head_html for analytics
 * introduce early app tour with Cicerone
+* introduce functioning ignore streets filter